[Refactor] Rename examples to align with the repo
