+++ examples/gdn/example_cumsum.py
